+++ examples/simple_async.rs
@@ -84,7 +84,12 @@ impl AsyncComponent for App {
         AsyncComponentParts { model, widgets }
     }
 
-    async fn update(&mut self, msg: Self::Input, _sender: AsyncComponentSender<Self>) {
+    async fn update(
+        &mut self,
+        msg: Self::Input,
+        _sender: AsyncComponentSender<Self>,
+        _root: &Self::Root,
+    ) {
         async_std::task::sleep(Duration::from_secs(1)).await;
         match msg {
             Msg::Increment => {